refactored add session
